Remove categorization compatibility exports
