+++ src/providers/BankTransactionsCategorizationStore/BankTransactionsCategorizationStoreProvider.tsx
@@ -1,4 +1,4 @@
-import { createContext, type PropsWithChildren, useContext, useMemo, useState } from 'react'
+import { createContext, type PropsWithChildren, useContext, useState } from 'react'
 import { createStore, useStore } from 'zustand'
 
 import type { BankTransactionCategoryComboBoxOption } from '@components/BankTransactionCategoryComboBox/bankTransactionCategoryComboBoxOption'
@@ -162,125 +162,6 @@ export function useGetAllBankTransactionsCategorizations(): { categorizations: M
   const categorizations = useStore(store, state => state.categorizations)
 
   return { categorizations }
-}
-
-type BankTransactionsCategoryActions = {
-  setTransactionCategory: (id: string, category: BankTransactionCategoryComboBoxOption | null) => void
-  setOnlyNewTransactionCategories: (transactionCategories: Array<{ id: string, category: BankTransactionCategoryComboBoxOption | null }>) => void
-  clearTransactionCategory: (id: string) => void
-  clearMultipleTransactionCategories: (ids: string[]) => void
-  clearAllTransactionCategories: () => void
-}
-
-export function useBankTransactionsCategoryActions(): BankTransactionsCategoryActions {
-  const {
-    setTransactionCategorization,
-    setOnlyNewTransactionCategorizations,
-  } = useBankTransactionsCategorizationActions()
-  const { categorizations } = useGetAllBankTransactionsCategorizations()
-
-  return useMemo(() => ({
-    setTransactionCategory: (id, category) => {
-      setTransactionCategorization(id, { category })
-    },
-    setOnlyNewTransactionCategories: (transactionCategories) => {
-      setOnlyNewTransactionCategorizations(new Map(
-        transactionCategories.map(({ id, category }) => [id, { category, taxCode: null }]),
-      ))
-    },
-    clearTransactionCategory: (id) => {
-      setTransactionCategorization(id, { category: null })
-    },
-    clearMultipleTransactionCategories: (ids) => {
-      ids.forEach((id) => {
-        setTransactionCategorization(id, { category: null })
-      })
-    },
-    clearAllTransactionCategories: () => {
-      categorizations.forEach((_, id) => {
-        setTransactionCategorization(id, { category: null })
-      })
-    },
-  }), [
-    categorizations,
-    setOnlyNewTransactionCategorizations,
-    setTransactionCategorization,
-  ])
-}
-
-export function useGetBankTransactionCategory(transactionId: string): { selectedCategory: BankTransactionCategoryComboBoxOption | null | undefined } {
-  const { selectedCategorization } = useGetBankTransactionCategorization(transactionId)
-
-  return {
-    selectedCategory: selectedCategorization?.category,
-  }
-}
-
-export function useGetAllBankTransactionsCategories(): { transactionCategories: Map<string, BankTransactionCategoryComboBoxOption | null> } {
-  const { categorizations } = useGetAllBankTransactionsCategorizations()
-  const transactionCategories = new Map<string, BankTransactionCategoryComboBoxOption | null>()
-
-  categorizations.forEach((categorization, id) => {
-    transactionCategories.set(id, categorization.category)
-  })
-
-  return {
-    transactionCategories,
-  }
-}
-
-type BankTransactionsTaxCodeActions = {
-  setTransactionTaxCode: (id: string, taxCode: TaxCodeSelectOption | null) => void
-  clearTransactionTaxCode: (id: string) => void
-  clearMultipleTransactionTaxCodes: (ids: string[]) => void
-  clearAllTransactionTaxCodes: () => void
-}
-
-export function useBankTransactionsTaxCodeActions(): BankTransactionsTaxCodeActions {
-  const {
-    setTransactionCategorization,
-    clearAllTransactionTaxCodes,
-  } = useBankTransactionsCategorizationActions()
-
-  return useMemo(() => ({
-    setTransactionTaxCode: (id, taxCode) => {
-      setTransactionCategorization(id, { taxCode })
-    },
-    clearTransactionTaxCode: (id) => {
-      setTransactionCategorization(id, { taxCode: null })
-    },
-    clearMultipleTransactionTaxCodes: (ids) => {
-      ids.forEach((id) => {
-        setTransactionCategorization(id, { taxCode: null })
-      })
-    },
-    clearAllTransactionTaxCodes,
-  }), [
-    clearAllTransactionTaxCodes,
-    setTransactionCategorization,
-  ])
-}
-
-export function useGetBankTransactionTaxCode(transactionId: string): {
-  selectedTaxCode: TaxCodeSelectOption | null
-  hasSelectedTaxCode: boolean
-} {
-  const { selectedCategorization } = useGetBankTransactionCategorization(transactionId)
-
-  return {
-    selectedTaxCode: selectedCategorization?.taxCode ?? null,
-    hasSelectedTaxCode: selectedCategorization !== undefined,
-  }
-}
-
-export function useGetBankTransactionSessionTaxCodes(transactionId: string): {
-  sessionTaxCodes: TaxCodeSelectOption[]
-} {
-  const { selectedTaxCode, hasSelectedTaxCode } = useGetBankTransactionTaxCode(transactionId)
-
-  return {
-    sessionTaxCodes: hasSelectedTaxCode && selectedTaxCode ? [selectedTaxCode] : [],
-  }
 }
 
 type BankTransactionsCategorizationStoreProviderProps = PropsWithChildren
